% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rle2.R
\name{rle2}
\alias{rle2}
\title{Run Length Encoding with Modifed \code{NA} handling}
\usage{
rle2(x)
}
\arguments{
\item{x}{an atomic vector (not a list)}
}
\value{
\code{rle2()} returns an object of class \code{"rle"} which is a list
with components:
  \item{lengths}{an integer vector containing the length of each run.}
\item{values}{a vector of the same length as \code{lengths} with the
  corresponding values.}
}
\description{
Compute the lengths and values of runs of equal values in a vector.
}
\details{
This is a drop-in replacement for \code{base::rle()} that treats all NAs as
identical. The returned object is the same as that produced by \code{base::rle()}
and is compatible with \code{base::inverse.rle()}.

This function is implemented in pure R - modelled after \code{base::rle()} and
should be compatible with whatever input the base function accepts.

If the input is a raw, logical, integer, numeric or character vector
then \code{rlenc()} may be a faster alternative.
}
