% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlenc.R
\name{rlenc}
\alias{rlenc}
\title{Fast run-length encoding.}
\usage{
rlenc(x)
}
\arguments{
\item{x}{atomic vector of raw, logical, integer, numeric or character data}
}
\value{
a data.frame with lengths, values and start location of each run
        within the given vector
}
\description{
Fast run-length encoding.
}
