% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlenc_compat.R
\name{rlenc_compat}
\alias{rlenc_compat}
\title{Run Length Encoding with Modifed \code{NA} handling}
\usage{
rlenc_compat(x)
}
\arguments{
\item{x}{an atomic vector (not a list)}
}
\value{
\code{rlenc_compat()} returns an object of class \code{"rle"} which
is a data.frame with components:
\item{start}{index of each first element in a run of a values}
\item{lengths}{an integer vector containing the length of each run.}
\item{values}{a vector of the same length as \code{lengths} with the
  corresponding values.}
}
\description{
Compute the lengths and values of runs of equal values in a vector.
}
\details{
This is a drop-in replacement for \code{base::rle()} that treats all NAs as
identical. The returned object is a data.frame which is compatible with
\code{base::inverse.rle()}.

This function is implemented in pure R - modelled after \code{base::rle()} and
should be compatible with every input the base function accepts.

If the input is a simple raw, logical, integer, numeric or character vector
then \code{rlenc()} may be a faster alternative.
}
