% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlenc.R
\name{rlenc_id}
\alias{rlenc_id}
\title{Fast run-length encoding of group id}
\usage{
rlenc_id(x)
}
\arguments{
\item{x}{atomic vector of raw, logical, integer, numeric or character data}
}
\value{
integer vector with group IDs associated with each run of identical
        values
}
\description{
Fast run-length encoding of group id
}
